feat: unify ts code without event emitter working
