+++ src/Clipboard.ts
@@ -2,7 +2,7 @@ import {EmitterSubscription, Platform} from 'react-native';
 import NativeClipboard, {
   addListener,
   removeAllListeners,
-} from './NativeClipboard';
+} from './NativeClipboardModule';
 
 /**
  * `Clipboard` gives you an interface for setting and getting content from Clipboard on both iOS and Android
